feat: :art: ajust: state foods = null / added: component RenderRecipe
