+++ src/common/components/RecipeInProgress.jsx
@@ -3,20 +3,17 @@ import { useParams } from 'react-router-dom';
 import store, { addRecDetail, setLoading } from '../../context/store';
 import { getStorage, setStorage } from '../../functions';
 import { fetchAPI, FETCH_ID_D, FETCH_ID_M } from '../../services';
-import ShareLikeButton from './ShareLikeButton';
+import RenderRecipe from './RenderRecipe';
 
 export default function RecipeInProgress() {
   const { id } = useParams();
-  // const [saveLS, setSaveLS] = useState(false);
   const [ingrOK, setIngrOK] = useState(() => (getStorage('inProgressRecipes')[id] || []));
   const [taskOK, setTaskOK] = useState({});
   const { recipes: { loading, recipeDetail, foods }, setRecipes } = useContext(store);
 
-  console.log(ingrOK);
-
   const addTaskCompleted = ({ target: { checked, name } }) => {
     const setTaskCompleted = { ...taskOK, [name]: checked };
-    console.log(setTaskCompleted);
+
     setTaskOK(setTaskCompleted);
 
     const ingredientsOK = Object.keys(setTaskCompleted)
@@ -56,6 +53,7 @@ export default function RecipeInProgress() {
                 data-testid={ `${i + 1}-ingredient-step` }
                 name={ task }
                 id={ ingredient }
+                checked={ ingrOK.includes(task) }
                 onClick={ addTaskCompleted }
               />
               {task}
@@ -66,47 +64,10 @@ export default function RecipeInProgress() {
     );
   };
 
-  function renderRecipe() {
-    return (
-      <div>
-        <img
-          data-testid="recipe-photo"
-          src={ recipeDetail.strMealThumb || recipeDetail.strDrinkThumb }
-          alt="recipe-img"
-          width="350px"
-        />
-        <div>
-          <div className="titleButtons">
-            <h1 data-testid="recipe-title">
-              { recipeDetail.strMeal || recipeDetail.strDrink }
-            </h1>
-            <ShareLikeButton />
-          </div>
-          <h5 data-testid="recipe-category">
-            Categoria:
-            { recipeDetail.strCategory }
-          </h5>
-        </div>
-        <div>
-          <h4>Ingredientes</h4>
-          {renderIngredients()}
-        </div>
-        <div>
-          <h4>Instruções</h4>
-          <p data-testid="instructions">{ recipeDetail.strInstructions }</p>
-        </div>
-        <button
-          data-testid="finish-recipe-btn"
-          type="button"
-        >
-          Finalizar Receita
-        </button>
-      </div>
-    );
-  }
-
   const getRecipeDetailByID = async () => {
-    if (foods) {
+    if (foods === null) {
+      setRecipes(setLoading(true));
+    } else if (foods) {
       const mealsDetails = await fetchAPI(`${FETCH_ID_M}${id}`);
       setRecipes(addRecDetail(mealsDetails.meals[0]));
       setRecipes(setLoading(false));
@@ -128,6 +89,6 @@ export default function RecipeInProgress() {
 
   if (loading) return (<h5>Loading...</h5>);
   return (
-    renderRecipe()
+    <RenderRecipe renderIngredients={ renderIngredients } />
   );
 }
